Fold SessionMetricsAggregator into types.py and Session
Per review feedback: the aggregator class and separate session_metrics
module diverged from the pattern already used for context-token metrics
(dataclass in types.py, aggregation method on Session, apply_* method
on RLMMetrics). This consolidates to the same shape:

- Move ProgrammaticToolCallStats and ChildSessionAggregate into
  types.py (all dataclasses in one place).
- ChildSessionAggregate owns the context-token sum/max key lists plus
  an absorb() method; the aggregation semantics live with the data.
- Move the child-walk back onto Session.aggregate_child_metrics, where
  it originally lived in the context-token branch.
- Add RLMMetrics.apply_programmatic_tool_call_stats(direct, child),
  mirroring the existing apply_child_aggregates. Session.finalize no
  longer assigns fields one-by-one.
- Delete src/rlm/session_metrics.py.

Output of meta.json, RLMMetrics field values, and aggregation semantics
are all unchanged — verified by the same end-to-end test.
