+++ src/rlm/types.py
@@ -2,6 +2,7 @@
 
 from __future__ import annotations
 
+import json
 from collections import deque
 from dataclasses import asdict, dataclass, field
 from pathlib import Path
@@ -38,6 +39,134 @@ class SummarizeApplied:
 
 
 BuiltinMetricEvent = IpythonExecuted | SummarizeRejected | SummarizeApplied
+
+
+@dataclass
+class ProgrammaticToolCallStats:
+    """Programmatic tool-call counts (skill CLIs invoked from the ipython REPL)."""
+
+    python_total: int = 0
+    bash_total: int = 0
+    by_tool_python: dict[str, int] = field(default_factory=dict)
+    by_tool_bash: dict[str, int] = field(default_factory=dict)
+
+    @classmethod
+    def from_log(cls, log_path: Path) -> ProgrammaticToolCallStats:
+        """Count programmatic tool calls from a session-local JSONL log.
+
+        Untrusted input (written by child processes that may exit mid-line),
+        so malformed entries are skipped.
+        """
+        stats = cls()
+        try:
+            f = open(log_path)
+        except FileNotFoundError:
+            return stats
+        with f:
+            for line in f:
+                try:
+                    entry = json.loads(line)
+                except json.JSONDecodeError:
+                    continue
+                tool = entry.get("tool")
+                source = entry.get("source")
+                if not isinstance(tool, str) or source not in {"python", "bash"}:
+                    continue
+
+                if source == "python":
+                    stats.python_total += 1
+                    stats.by_tool_python[tool] = stats.by_tool_python.get(tool, 0) + 1
+                else:
+                    stats.bash_total += 1
+                    stats.by_tool_bash[tool] = stats.by_tool_bash.get(tool, 0) + 1
+
+        return stats
+
+    @classmethod
+    def from_meta(cls, meta: dict) -> ProgrammaticToolCallStats:
+        """Load tool-call stats previously persisted via ``to_dict``."""
+        raw = meta.get("programmatic_tool_call_stats", {})
+        return cls(
+            python_total=int(raw.get("python_total", 0)),
+            bash_total=int(raw.get("bash_total", 0)),
+            by_tool_python=dict(raw.get("by_tool_python", {})),
+            by_tool_bash=dict(raw.get("by_tool_bash", {})),
+        )
+
+    def merge(self, other: ProgrammaticToolCallStats) -> ProgrammaticToolCallStats:
+        """Return a merged copy of this stats object and *other*."""
+        merged = ProgrammaticToolCallStats(
+            python_total=self.python_total + other.python_total,
+            bash_total=self.bash_total + other.bash_total,
+            by_tool_python=self.by_tool_python.copy(),
+            by_tool_bash=self.by_tool_bash.copy(),
+        )
+        for tool_name, count in other.by_tool_python.items():
+            merged.by_tool_python[tool_name] = (
+                merged.by_tool_python.get(tool_name, 0) + count
+            )
+        for tool_name, count in other.by_tool_bash.items():
+            merged.by_tool_bash[tool_name] = (
+                merged.by_tool_bash.get(tool_name, 0) + count
+            )
+        return merged
+
+    def to_dict(self) -> dict[str, Any]:
+        return asdict(self)
+
+
+def _empty_context_token_stats() -> dict[str, int]:
+    return {
+        "session_count": 0,
+        "input_tokens_total": 0,
+        "output_tokens_total": 0,
+        "final_input_tokens_total": 0,
+        "final_output_tokens_total": 0,
+        "branch_count": 0,
+        "branch_input_tokens_sum": 0,
+        "branch_input_tokens_max": 0,
+        "branch_output_tokens_sum": 0,
+        "branch_output_tokens_max": 0,
+    }
+
+
+@dataclass
+class ChildSessionAggregate:
+    """Aggregated stats across all recursive descendants of a session."""
+
+    context_token_stats: dict[str, int] = field(
+        default_factory=_empty_context_token_stats
+    )
+    tool_call_stats: ProgrammaticToolCallStats = field(
+        default_factory=ProgrammaticToolCallStats
+    )
+
+    _CONTEXT_TOKEN_SUM_KEYS = (
+        "session_count",
+        "input_tokens_total",
+        "output_tokens_total",
+        "final_input_tokens_total",
+        "final_output_tokens_total",
+        "branch_count",
+        "branch_input_tokens_sum",
+        "branch_output_tokens_sum",
+    )
+    _CONTEXT_TOKEN_MAX_KEYS = ("branch_input_tokens_max", "branch_output_tokens_max")
+
+    def absorb(
+        self,
+        ctx_stats: dict[str, int],
+        tool_stats: ProgrammaticToolCallStats,
+    ) -> None:
+        """Combine a child session's stats into this aggregate."""
+        for key in self._CONTEXT_TOKEN_SUM_KEYS:
+            self.context_token_stats[key] += int(ctx_stats.get(key, 0))
+        for key in self._CONTEXT_TOKEN_MAX_KEYS:
+            self.context_token_stats[key] = max(
+                self.context_token_stats[key],
+                int(ctx_stats.get(key, 0)),
+            )
+        self.tool_call_stats = self.tool_call_stats.merge(tool_stats)
 
 
 @dataclass
@@ -178,6 +307,16 @@ class RLMMetrics:
             "branch_output_tokens_max", 0
         )
         self._refresh_derived_metrics()
+
+    def apply_programmatic_tool_call_stats(
+        self,
+        direct: ProgrammaticToolCallStats,
+        child: ProgrammaticToolCallStats,
+    ) -> None:
+        self.programmatic_tool_calls_python = direct.python_total
+        self.programmatic_tool_calls_bash = direct.bash_total
+        self.sub_rlm_programmatic_tool_calls_python = child.python_total
+        self.sub_rlm_programmatic_tool_calls_bash = child.bash_total
 
     def context_token_stats(self) -> dict[str, int]:
         self._refresh_derived_metrics()
